refactor(ui): TasksPanel + /jobs/:id/tasks; task detail renders through the shell
