+++ frontend/src/App.svelte
@@ -79,10 +79,11 @@
     '/jobs/:id': JobDetailPage,
     '/jobs/:id/edit': JobEditPage,
     '/jobs/:id/duplicate': DuplicateJobPage,
-    '/jobs/:id/tasklist': JobTaskListPage,
     '/jobs/:id/history': JobHistoryPage,
     '/jobs/:jobId/shipments': JobShipmentsPage,
+    '/jobs/:jobId/tasks': JobTaskListPage,
     '/jobs/:jobId/tasks/:taskId': TaskDetailPage,
+    '/jobs/:id/tasklist': JobTaskListPage,
     '/shipments/:sid/print': PackingListPrint,
     '/estimates/:id/wizard': EstimateWizardPage,
     '/estimates/:id/send': EstimateSendPage,
